+++ src/edit-param-guard/constants.ts
@@ -1,0 +1,3 @@
+export const LOG_PREFIX = "[dingtalk-connector][edit-param-guard]"
+
+export const EDIT_TOOL_NAME = "edit"